refactor(step2): LottoNumber Cache 적용 및 Ticket에 생성자 추가
