+++ src/main/kotlin/lotto/LottoApplication.kt
@@ -1,5 +1,9 @@
 package lotto
 
+import lotto.domain.EarnRatio
+import lotto.domain.LottoNumber
+import lotto.domain.LottoShop
+import lotto.domain.LottoTicketGenerator
 import lotto.view.getLuckyNumbers
 import lotto.view.getMoneyForTickets
 import lotto.view.printEarnRatio
@@ -7,6 +11,8 @@ import lotto.view.printLottoResults
 import lotto.view.printLottoTickets
 
 fun main() {
+    println(LottoNumber.of(5))
+
     val lottoShop = LottoShop(LottoTicketGenerator())
     val inputMoney = getMoneyForTickets()
     val lottoTickets = lottoShop.getAutoTickets(inputMoney)
