+++ src/main/kotlin/lotto/domain/EarnRatio.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 object EarnRatio {
