+++ src/main/kotlin/lotto/domain/LottoTicketGenerateStrategy.kt
@@ -1,4 +1,6 @@
-package lotto
+package lotto.domain
+
+import lotto.domain.LottoTicket
 
 interface LottoTicketGenerateStrategy {
