+++ src/main/kotlin/lotto/domain/LottoTicketGenerator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoTicketGenerator : LottoTicketGenerateStrategy {
