+++ src/test/kotlin/lotto/domain/LottoResultTest.kt
@@ -1,5 +1,6 @@
-package lotto
+package lotto.domain
 
+import lotto.domain.LottoResult
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
