+++ src/test/kotlin/lotto/domain/LottoShopTest.kt
@@ -1,5 +1,11 @@
-package lotto
+package lotto.domain
 
+import lotto.domain.LottoNumber
+import lotto.domain.LottoResult
+import lotto.domain.LottoShop
+import lotto.domain.LottoTicket
+import lotto.domain.LottoTicketGenerateStrategy
+import lotto.domain.LottoTicketGenerator
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
